refact: propagate refund ignoreBalanceCheck param to payment providers
